+++ src/Pdf/DompdfRenderer.php
@@ -2,6 +2,7 @@
 
 namespace OccTherapist\AdvancedTableExportForFilament\Pdf;
 
+use Closure;
 use Dompdf\Dompdf;
 use OccTherapist\AdvancedTableExportForFilament\Contracts\PdfRenderer;
 use RuntimeException;
@@ -17,6 +18,11 @@ class DompdfRenderer implements PdfRenderer
         $dompdf = new Dompdf;
         $dompdf->loadHtml($html);
         $dompdf->setPaper('A4', ($options['orientation'] ?? 'landscape') === 'landscape' ? 'landscape' : 'portrait');
+
+        if (isset($options['modifyDompdfWriter']) && $options['modifyDompdfWriter'] instanceof Closure) {
+            ($options['modifyDompdfWriter'])($dompdf, $options['context'] ?? []);
+        }
+
         $dompdf->render();
 
         return base64_encode($dompdf->output());